fix(build): stop the client bundle from reaching server-only modules, and make the guard find them
