+++ open-sse/utils/cursorAgentCliVersionPin.ts
@@ -1,0 +1,7 @@
+// Import-free pin so client bundles can read the version without pulling node-only detection code.
+
+/**
+ * Pinned Agent CLI build id used when no local install is found (typical
+ * headless OmniRoute). Bump when refreshing Cursor CLI impersonation.
+ */
+export const CURSOR_AGENT_CLI_VERSION = "2026.07.08-0c04a8a";
